Don't make edge_vec_check ints

+++ src/mattertune/backbones/orb/model.py
@@ -634,7 +634,7 @@ class ORBBackboneModule(
                 N_paths += len(paths_info[struct_i][j]["Reffs"])
 
             edge_match = torch.zeros((N_paths), dtype = torch.int64, device = device)
-            edge_vec_check = torch.zeros((N_paths, 3), dtype = torch.int64, device = device)
+            edge_vec_check = torch.zeros((N_paths, 3), device = device)
             match_failed = torch.zeros((N_paths), dtype = torch.int64, device = device)
             tot_path = 0
             for j, abs_i in enumerate(abs_inds[struct_i]):
